Include SmtpClient._transport.Credentials in tests, move tests to own file
